@@ -20,7 +20,11 @@
         "resolve-url-loader": "^2.3.1",
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
-        "vue": "^2.5.17",
-        "vue-template-compiler": "^2.6.8"
+        "vue": "^2.6.8",
+        "vue-template-compiler": "^2.6.9",
+        "vue-i18n": "^8.9.0",
+        "vue-meta": "^1.5.8",
+        "vue-router": "^3.0.2",
+        "vuex": "^3.1.0"
     }
 }
